backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-dotenv
psycopg2-binary
supabase